Don't close the previous connection if the current connection fails.
The bug here is basically this:

    while true
      begin
        foo = server.accept
      rescue => e
        foo.close rescue nil
      end
    end

If `server.accept` throws an exception, `foo` will never get assigned.
This means `foo` will always only be `nil` or the previously-accepted
connection:

1) connection 1: foo = server.accept  #success!
2) connection 2: foo = server.accept # fails to assign if a TLS handshake fails
3) connection 2 error: foo.close # calls `foo.close` from item 1 (the
   first connection, not the 2nd!)

I've done some tests, and it seems like OpenSSL::SSL::SSLSocket#close
will close the underlying IO (TCPSocket) on `accept` failures anyway, so
we don't need to do this.

That was a super weird thing to debug, haha.

Fixes some of #34

Fixes #43