Make this input plugin capable of doing line delimiting.
This change keeps the behavior the same, but inverts the previous
implementation which used `fix_streaming_codecs` to automatically
switch from `plain` to `lines` and from `json` to `json_lines`. The
switch was needed because the input wasn't programmed to make
choices about line delimiters -- that job was handled by the `line` and
`json_line` codec.

This particular change was driven by a discussion about improving the
`cef` codec and how I personally didn't want to have both a `cef` and
`cef_lines` codec.

I've had discussions over the past year or two with folks on the
Logstash team where we talked about separating "payload delimiting"
concerns away from the codecs and moving that delimiting functionality
into a new system we tentatively called a "Mill" (credit to @guyboertje
for the term and I think the original idea) after the idea of a sawmill
which cuts raw material (logs) into usable pieces (lumber).

This commit doesn't introduce "mill" as a named concept, but implements
the idea that codecs are probably not the best place for delimiting to
happen. You can feel the weirdness of some codecs doing delimiting in
several places:

* The file input (files are a stream of bytes) emitting *lines* to its codecs.
* The tcp input emitting *bytes* to its codecs
* The fact that `json` and `json_lines` codecs exist and do the same
thing except one does line delimiting also.
* The fact that `plain` and `line` codecs exist and do the same thing
except one does line delimiting also.
* The `fix_streaming_codecs` method that hacks around the previous two
points to force (previously) the `line` codec to be used on the tcp
input if the user mistakenly configured `plain`. Same for JSON.