Reduce instances where we're passing around socket
Reducing the number of times we refer to the socket object so that
most of the interaction lives in the `handle_socket` method.

Remove redundant conditional in ssl_subject assignment, it's already
assumed that we're in client mode at that point.

Necessary to also fix up refderences in the DecoderImpl class as well,
since the tcp plugin's decode_buffer method changed.